feat: Tanstack Query 기반 서버 상태 관리 적용

+++ src/components/todo-list/todo-item.tsx
@@ -1,21 +1,51 @@
 import { Button } from "@/components/ui/button";
-import { useDeleteTodo } from "@/store/todos";
-import type { Todo } from "@/types/todo-type";
+import { Checkbox } from "@/components/ui/checkbox";
+import {
+  useDeleteTodo,
+  useTodoDataById,
+  useUpdateTodo,
+} from "@/hooks/queries/use-todo-list";
+import { Link } from "react-router";
 
-export default function TodoItem({ id, content }: Todo) {
-  const deleteTodo = useDeleteTodo();
+export default function TodoItem({ id }: { id: string }) {
+  const { data: todo, isLoading } = useTodoDataById(id, "LIST");
+  if (isLoading) return <div>로딩중...</div>;
+  if (!todo) return;
+  const { content, isDone } = todo;
 
-  const handleDeleteItem = () => {
+  const { mutate: updateTodo } = useUpdateTodo();
+  const { mutate: deleteTodo, isPending: isDeleting } = useDeleteTodo();
+
+  const handleClickCheckbox = () => {
+    updateTodo({
+      id,
+      isDone: !isDone,
+    });
+  };
+
+  const handleClickDelete = () => {
+    const confirmed = window.confirm("정말로 삭제하시겠습니까?");
+    if (!confirmed) return;
     deleteTodo(id);
   };
 
   return (
     <div
       id={"todo" + id}
-      className="radius flex items-center justify-between rounded-md border p-2 pl-3"
+      className="radius flex items-center justify-start gap-2 rounded-md border p-2 pl-3 hover:bg-gray-50"
     >
-      {content}
-      <Button variant="destructive" onClick={handleDeleteItem}>
+      <Checkbox
+        checked={isDone}
+        onClick={handleClickCheckbox}
+        disabled={isDeleting}
+      />
+      <Link to={`/todolist/${id}`}>{content}</Link>
+      <Button
+        variant="destructive"
+        onClick={handleClickDelete}
+        className="ml-auto"
+        disabled={isDeleting}
+      >
         삭제
       </Button>
     </div>
